requirements.txt
Django==2.1.5
django-autocomplete-light==3.3.2
django-bootstrap3==11.0.0
django-filter==2.0.0
django-tables2==2.0.3
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
